requirements.txt
FLASK
python-dotenv
sqlalchemy
sqlalchemy-orm
psycopg2